Change AuthStatusContext from .js to .ts
