+++ tslint.json
@@ -9,7 +9,8 @@
     "rules": {
       "interface-name": [true, "never-prefix"],
       "semicolon": [true, "never"],
-      "quotemark": [true, "single"]
+      "quotemark": [true, "single"],
+      "member-access": false
     },
     "rulesDirectory": []
 }
